# route/v1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is msp_backend api list.",
        "title": "MSP_BACKEND API",
        "contact": {},
        "license": {},
        "version": "1.0.0"
    },
    "host": "192.168.2.80:8100",
    "basePath": "/api/v1",
    "paths": {
...
        "/log_platform": {
            "get": {
                "description": "获取所有接入日至平台的服务列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "日志平台"
                ],
                "summary": "Get LogPlatformDataList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "第几页，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页多少条，默认10",
                        "name": "limit_num",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2
                        ],
                        "type": "integer",
                        "description": "是否关闭流量，1开启2关闭",
                        "name": "is_block",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            },
            "put": {
                "description": "修改日志平台数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "日志平台"
                ],
                "summary": "Edit LogPlatformData",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "主键ID",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "服务名称",
                        "name": "service_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "服务appid",
                        "name": "appid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "日志log_key",
                        "name": "log_key",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            1,
                            2
                        ],
                        "type": "integer",
                        "description": "流量开关，1开2关",
                        "name": "is_block",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "速率限制",
                        "name": "rate_limit",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            },
            "post": {
                "description": "新增日志平台数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "日志平台"
                ],
                "summary": "Add LogPlatformData",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务名称",
                        "name": "service_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "服务appid",
                        "name": "appid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "日志log_key",
                        "name": "log_key",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            1,
                            2
                        ],
                        "type": "integer",
                        "description": "流量开关，1开2关",
                        "name": "is_block",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "速率限制",
                        "name": "rate_limit",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            }
        },
        "/log_platform/count": {
            "get": {
                "description": "获取日志平台key的数量",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "日志平台"
                ],
                "summary": "Get LogKeyCount",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "是否关闭流量，2关闭1开启",
                        "name": "is_block",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "速率限制，单位：次/s",
                        "name": "rate_limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            }
        },
...
        "/log_platform/info/{id}": {
            "get": {
                "description": "获取日志平台某个服务详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "日志平台"
                ],
                "summary": "Get LogPlatformDataInfo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务主键ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            }
        },
...
        "/log_platform/{id}": {
            "delete": {
                "description": "删除日志平台某个服务",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "日志平台"
                ],
                "summary": "Delete LogPlatformData",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务主键ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Res"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.Res": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}